--- a/.gitignore
+++ b/.gitignore
@@ -9,3 +9,4 @@ erl_crash.dump
 *#
 #*
 *~
+/bench/snapshots