--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -1,0 +1,5 @@
+###Improve Money.make performance for:
+  * `from_integer`
+  * `from_decimal`
+  * `from_float`
+  * `zero`