--- a/mix.exs
+++ b/mix.exs
@@ -25,7 +25,7 @@ defmodule Monetized.Mixfile do
       {:inch_ex, "~> 0.5.1",  only: :docs},
       {:decimal, "~> 1.1.2"},
       {:ecto,    "~> 1.1.7"},
-      {:benchfella, "~> 0.3.2", only: :dev},
+      {:benchfella, "~> 0.3.2", only: :bench},
     ]
   end