Improve performance of `from_decimal`, `from_integer`, `from_float`. These are microbenchmarks so to be taken w/ a grain of salt, but I think you can see that in the implementation that circumventing doing the `to_string` dance would improve performance

before
```
make with formatted decimal               100000   12.78 µs/op
make with unformatted decimal             100000   17.89 µs/op
make with string                           50000   35.91 µs/op
make with string and currency option       50000   41.40 µs/op
make with integer                          50000   42.34 µs/op
zero                                       50000   42.46 µs/op
make with float                            50000   42.55 µs/op
```
After
```
zero                                     1000000   1.16 µs/op
make with formatted decimal              1000000   1.23 µs/op
make with integer                        1000000   1.38 µs/op
make with float                           500000   4.62 µs/op
make with unformatted decimal             100000   18.11 µs/op
make with string                          100000   23.50 µs/op
make with string and currency option       50000   32.78 µs/op
```